+++ tests/helpers.py
@@ -1,5 +1,6 @@
 from typing import List, Type
 
+from click.testing import Result
 from pydantic import BaseSettings
 from pytest_mock import MockerFixture
 from typer.testing import CliRunner
@@ -8,13 +9,21 @@ from settings_doc.main import app
 
 
 def run_app_with_settings(
-    mocker: MockerFixture, runner: CliRunner, settings: Type[BaseSettings], args: List[str] = None
+    mocker: MockerFixture,
+    runner: CliRunner,
+    settings: Type[BaseSettings],
+    args: List[str] = None,
+    fmt: str = "markdown",
 ) -> str:
     if args is None:
         args = []
 
     mocker.patch("settings_doc.main.import_class_path", return_value=settings)
     result = runner.invoke(
-        app, ["generate", "--class", "MockedClass", "--output-format", "markdown"] + args, catch_exceptions=False
+        app, ["generate", "--class", "MockedClass", "--output-format", fmt] + args, catch_exceptions=False
     )
     return result.stdout.lower()
+
+
+def copy_templates(runner: CliRunner, folder: str) -> Result:
+    return runner.invoke(app, ["templates", "--copy-to", folder], catch_exceptions=False)