+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Types of changes are:
 
 ## [Unreleased]
 
+### Fixed
+
+- Description of `generate --templates`.
+
 ## [0.4.0] - 2021-11-11
 
 ### Changed